Alphazero

+++ aithena/util/dirichlet.cc
@@ -16,7 +16,7 @@ template <class RNG>
 void dirichlet_distribution<RNG>::set_params(const std::vector<double>& new_params) {
   alpha = new_params;
   std::vector<std::gamma_distribution<>> new_gamma(alpha.size());
-  for (int i = 0; i < alpha.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(alpha.size()); ++i) {
     std::gamma_distribution<> temp(alpha[i], 1);
     new_gamma[i] = temp;
   }
@@ -30,9 +30,9 @@ std::vector<double> dirichlet_distribution<RNG>::get_params() {
 
 template <class RNG>
 std::vector<double> dirichlet_distribution<RNG>::operator()(RNG& generator) {
-  std::vector<double> x(alpha.size());
+  std::vector<double> x(static_cast<int>(alpha.size()));
   double sum = 0.0;
-  for (int i = 0; i < alpha.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(alpha.size()); ++i) {
     x[i] = gamma[i](generator);
     sum += x[i];
   }
